pyproject.toml
[project]
name = "bluebird"
version = "2.0.0"
description = "Monitor users on X and report new posts via Discord."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "environs>=14.1.1",
    "loguru>=0.7.3",
    "loguru-discord>=1.4.0",
    "httpx>=0.28.1",
    "discord-clyde>=0.2.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "ruff>=0.11.10",
]

[tool.ruff]
fix = true
show-fixes = true

[tool.ruff.lint]
select = ["I"]

[tool.ruff.lint.isort]
split-on-trailing-comma = false
